// Zybach.Web/src/app/shared/generated/enum/field-definition-type-enum.ts
//  IMPORTANT:
//  This file is generated. Your changes will be lost.
//  Source Table: [dbo].[FieldDefinitionType]

export enum FieldDefinitionTypeEnum {
  Name = 1,
  HasWaterLevelInspections = 2,
  HasWaterQualityInspections = 3,
  LatestWaterLevelInspectionDate = 4,
  LatestWaterQualityInspectionDate = 5,
  WellRegistrationNumber = 6,
  WellNickname = 7,
  AgHubRegisteredUser = 8,
  WellFieldName = 9,
  IrrigationUnitID = 10,
  WellIrrigatedAcres = 11,
  WellChemigationInspectionParticipation = 12,
  WellWaterLevelInspectionParticipation = 13,
  WellWaterQualityInspectionParticipation = 14,
  WellProgramParticipation = 15,
  WellOwnerName = 16,
  SensorLastMessageAgeHours = 17,
  SensorLastVoltageReading = 18,
  SensorFirstReadingDate = 19,
  SensorLastReadingDate = 20,
  SensorStatus = 21,
  SensorType = 22,
  IrrigationUnitAcres = 23,
  SensorLastVoltageReadingDate = 24,
  SensorRetirementDate = 25,
  ContinuityMeterStatus = 26,
  ActiveSupportTicket = 27
}